# openapi.yaml
openapi: 3.0.1
info:
  title: Presidio Todos Test API
  version: 1.0.0
servers:
  - url: https://todo-api-1rid.onrender.com
    description: Development server
paths:
  /todos:
    get:
      summary: Get all todos
      operationId: getTodos
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
    post:
      summary: Create multiple todos
      operationId: createTodos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                todos:
                  type: array
                  items:
                    $ref: "#/components/schemas/Todo"
      responses:
        "201":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
  /todos/{id}:
    put:
      summary: Update a todo
      operationId: updateTodo
      parameters:
        - name: id
          in: path
          description: ID of todo to update
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        description: Todo object to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Todo"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
    delete:
      summary: Delete multiple todos by IDs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - title
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string